field_study_scraper/pyproject.toml
[tool.poetry]
name = "github_api"
version = "0.1.0"
description = ""
authors = ["Lukas Döllerer <lukas.doellerer@tum.de>"]

[tool.poetry.dependencies]
python = "^3.9"
pandas = "^1.3.4"
PyGithub = "^1.55"
PyDriller = "^2.0"
matplotlib = "^3.5.1"

[tool.poetry.dev-dependencies]
black = "^21.10b0"
ipython = "^7.29.0"

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
